refactored Connect Four Components into their own files
