+++ board-games/src/components/ConnectFour/components/ConnectFourBoard/ConnectFourBoard.module.css
@@ -1,0 +1,9 @@
+/* uses CSS grid to create a container for the board */
+.board { 
+    min-height: 600px;
+    min-width: 700px;
+    
+    display: grid;
+    grid-template-columns: repeat(7, 1fr); /* 7 columns */
+    grid-template-rows: repeat(6, 1fr); /* 6 rows */
+}
