+++ board-games/src/components/ConnectFour/components/ConnectFourBoardSpace/ConnectFourBoardSpace.module.css
@@ -1,0 +1,10 @@
+/* a single grid of the board */
+.boardSpace {
+    width: 100%;
+    height: 100%;
+    background-color: rgb(45, 102, 209);
+    
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
