+++ board-games/src/components/ConnectFour/components/index.js
@@ -1,5 +1,11 @@
+import ConnectFourBoard from "./ConnectFourBoard/ConnectFourBoard";
+import ConnectFourBoardSpace from "./ConnectFourBoardSpace/ConnectFourBoardSpace";
+import ConnectFourPiece from "./ConnectFourPiece/ConnectFourPiece";
 import Menu from "./Menu/Menu";
 
 export {
+    ConnectFourBoard,
+    ConnectFourBoardSpace,
+    ConnectFourPiece,
     Menu
 }
